+++ types/staking.go
@@ -25,12 +25,12 @@ const (
 var PowerReduction = NewIntFromBigInt(new(big.Int).Exp(big.NewInt(10), big.NewInt(6), nil))
 
 // TokensToConsensusPower - convert input tokens to potential consensus-engine power
-func TokensToConsensusPower(tokens Int) int64 {
+func TokensToConsensusPower(tokens BigInt) int64 {
 	return (tokens.Quo(PowerReduction)).Int64()
 }
 
 // TokensFromConsensusPower - convert input power to tokens
-func TokensFromConsensusPower(power int64) Int {
+func TokensFromConsensusPower(power int64) BigInt {
 	return NewInt(power).Mul(PowerReduction)
 }
