+++ x/gov/types/genesis.go
@@ -7,12 +7,12 @@ import (
 
 // GenesisState - all auth state that must be provided at genesis
 type GenesisState struct {
-	Params    Params  `json:"params" yaml:"params"`
-	DAOTokens sdk.Int `json:"DAO_Tokens"`
+	Params    Params     `json:"params" yaml:"params"`
+	DAOTokens sdk.BigInt `json:"DAO_Tokens"`
 }
 
 // NewGenesisState - Create a new genesis state
-func NewGenesisState(params Params, daoTokens sdk.Int) GenesisState {
+func NewGenesisState(params Params, daoTokens sdk.BigInt) GenesisState {
 	return GenesisState{
 		Params:    params,
 		DAOTokens: daoTokens,
