+++ x/nodes/exported/exported.go
@@ -14,6 +14,6 @@ type ValidatorI interface {
 	GetStatus() sdk.StakeStatus     // status of the validator
 	GetAddress() sdk.Address        // operator address to receive/return validators coins
 	GetPublicKey() crypto.PublicKey // validation consensus pubkey
-	GetTokens() sdk.Int             // validation tokens
+	GetTokens() sdk.BigInt          // validation tokens
 	GetConsensusPower() int64       // validation power in tendermint
 }
