+++ x/nodes/keeper/account.go
@@ -6,7 +6,7 @@ import (
 )
 
 // GetBalance - Retrieve balance for account
-func (k Keeper) GetBalance(ctx sdk.Ctx, addr sdk.Address) sdk.Int {
+func (k Keeper) GetBalance(ctx sdk.Ctx, addr sdk.Address) sdk.BigInt {
 	coins := k.AccountKeeper.GetCoins(ctx, addr)
 	return coins.AmountOf(k.StakeDenom(ctx))
 }
@@ -23,7 +23,7 @@ func (k Keeper) GetAccount(ctx sdk.Ctx, addr sdk.Address) (acc *auth.BaseAccount
 }
 
 // SendCoins - Deliver coins to account
-func (k Keeper) SendCoins(ctx sdk.Ctx, fromAddress sdk.Address, toAddress sdk.Address, amount sdk.Int) sdk.Error {
+func (k Keeper) SendCoins(ctx sdk.Ctx, fromAddress sdk.Address, toAddress sdk.Address, amount sdk.BigInt) sdk.Error {
 	coins := sdk.NewCoins(sdk.NewCoin(k.StakeDenom(ctx), amount))
 	err := k.AccountKeeper.SendCoins(ctx, fromAddress, toAddress, coins)
 	if err != nil {
