+++ x/pocketcore/keeper/pool.go
@@ -11,25 +11,25 @@ func (k Keeper) StakeDenom(ctx sdk.Ctx) (res string) {
 }
 
 // "GetAppStakedTokens" - Returns the total number of staked tokens in the apps module
-func (k Keeper) GetAppStakedTokens(ctx sdk.Ctx) (res sdk.Int) {
+func (k Keeper) GetAppStakedTokens(ctx sdk.Ctx) (res sdk.BigInt) {
 	res = k.appKeeper.GetStakedTokens(ctx)
 	return
 }
 
 // "GetNodeStakedTokens" - Returns the total number of staked tokens in the nodes module
-func (k Keeper) GetNodesStakedTokens(ctx sdk.Ctx) (res sdk.Int) {
+func (k Keeper) GetNodesStakedTokens(ctx sdk.Ctx) (res sdk.BigInt) {
 	res = k.posKeeper.GetStakedTokens(ctx)
 	return
 }
 
 // "GetTotalTokens" - Returns the total number of tokens kept in any/all modules
-func (k Keeper) GetTotalTokens(ctx sdk.Ctx) (res sdk.Int) {
+func (k Keeper) GetTotalTokens(ctx sdk.Ctx) (res sdk.BigInt) {
 	res = k.posKeeper.TotalTokens(ctx)
 	return
 }
 
 // "GetTotalStakedTokens" - Returns the summation of app staked tokens and node staked tokens
-func (k Keeper) GetTotalStakedTokens(ctx sdk.Ctx) (res sdk.Int) {
+func (k Keeper) GetTotalStakedTokens(ctx sdk.Ctx) (res sdk.BigInt) {
 	res = k.GetNodesStakedTokens(ctx).Add(k.GetAppStakedTokens(ctx))
 	return
 }
